refactor: Reorganize code and unit tests
Reorganized WMG API code and unit tests minimally to
enable writing unit tests for WMG V2 API in isolation
from the unit tests associated with WMG V1 API.
