+++ tests/unit/backend/wmg/api/test_v2.py
@@ -3,7 +3,7 @@ import unittest
 from unittest.mock import patch
 
 from backend.api_server.app import app
-from backend.wmg.api.v1 import find_dimension_id_from_compare
+from backend.wmg.api.v2 import find_dimension_id_from_compare
 from backend.wmg.data.query import MarkerGeneQueryCriteria
 from tests.unit.backend.fixtures.environment_setup import EnvironmentSetup
 from tests.unit.backend.wmg.fixtures.test_cube_schema import expression_summary_non_indexed_dims
@@ -93,7 +93,7 @@ def generate_expected_expression_summary_dictionary(genes, tissues, cell_types, 
 
 def generate_test_inputs_and_expected_outputs(genes, tissues, organism, dim_size, me, expected_count, compare_dim=None):
     """
-    Generates test inputs and expected outputs for the /wmg/v1/query endpoint.
+    Generates test inputs and expected outputs for the /wmg/v2/query endpoint.
 
     Arguments
     ---------
@@ -108,7 +108,7 @@ def generate_test_inputs_and_expected_outputs(genes, tissues, organism, dim_size
     Returns
     -------
     tuple of (request, expected_expression_summary, expected_term_id_labels)
-    request: dictionary containing the request body to send to the /wmg/v1/query endpoint
+    request: dictionary containing the request body to send to the /wmg/v2/query endpoint
     expected_expression_summary: dictionary containing the expected expression summary values
     expected_term_id_labels: dictionary containing the expected term ID labels
     """
@@ -156,7 +156,7 @@ def generate_test_inputs_and_expected_outputs(genes, tissues, organism, dim_size
     )
 
 
-class WmgApiV1Tests(unittest.TestCase):
+class WmgApiV2Tests(unittest.TestCase):
     """
     Tests WMG API endpoints. Tests the flask app only, and not other stack dependencies, such as S3. Builds and uses a
     temporary WMG cube on local filesystem to avoid dependency on localstack S3.
@@ -172,19 +172,19 @@ class WmgApiV1Tests(unittest.TestCase):
         super().setUpClass()
         cls.maxDiff = None
 
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__primary_filter_dimensions__returns_200(self, load_snapshot):
         # This test appears to be hitting a TileDB (<=0.13.1) bug and fails (intermittently) if dim_size=3
         with create_temp_wmg_snapshot(dim_size=1) as snapshot:
             # setup up API endpoints to use a mocked cube
             load_snapshot.return_value = snapshot
-            response = self.app.get("/wmg/v1/primary_filter_dimensions")
+            response = self.app.get("/wmg/v2/primary_filter_dimensions")
 
         self.assertEqual(200, response.status_code)
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__primary_filter_dimensions__returns_valid_response_body(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -201,7 +201,7 @@ class WmgApiV1Tests(unittest.TestCase):
             ontology_term_label.side_effect = lambda ontology_term_id: f"{ontology_term_id}_label"
             gene_term_label.side_effect = lambda gene_term_id: f"{gene_term_id}_label"
 
-            response = self.app.get("/wmg/v1/primary_filter_dimensions")
+            response = self.app.get("/wmg/v2/primary_filter_dimensions")
 
         expected = dict(
             snapshot_id=test_snapshot_id,
@@ -212,9 +212,9 @@ class WmgApiV1Tests(unittest.TestCase):
 
         self.assertEqual(expected, json.loads(response.data))
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_single_primary_dims__returns_200_and_correct_response(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -242,7 +242,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, 10
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
@@ -253,9 +253,9 @@ class WmgApiV1Tests(unittest.TestCase):
             }
             self.assertEqual(expected_response, json.loads(response.data))
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_single_tissue_no_genes__returns_200_and_correct_response(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -286,7 +286,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, 10
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
@@ -297,9 +297,9 @@ class WmgApiV1Tests(unittest.TestCase):
             }
             self.assertEqual(expected_response, json.loads(response.data))
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_multiple_tissues_no_genes__returns_200_and_correct_response(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -330,7 +330,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, 10
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
@@ -341,9 +341,9 @@ class WmgApiV1Tests(unittest.TestCase):
             }
             self.assertEqual(expected_response, json.loads(response.data))
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_request_multi_primary_dims_only__returns_200_and_correct_response(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -371,7 +371,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, 10
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
@@ -382,9 +382,9 @@ class WmgApiV1Tests(unittest.TestCase):
             }
             self.assertEqual(expected, json.loads(response.data))
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_request_multi_primary_dims_only_with_compare__returns_200_and_correct_response(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -412,7 +412,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, 10, compare_dim="self_reported_ethnicity"
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
@@ -423,9 +423,9 @@ class WmgApiV1Tests(unittest.TestCase):
             }
             self.assertEqual(expected, json.loads(response.data))
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_explicit_cell_ordering__returns_correct_cell_ordering(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -454,16 +454,16 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, 10
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
             expected = expected_term_id_labels["cell_types"]
             self.assertEqual(expected, json.loads(response.data)["term_id_labels"]["cell_types"])
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_total_cell_count_per_cell_type(self, load_snapshot, ontology_term_label, gene_term_label):
         expected_count = 42
         dim_size = 2
@@ -491,7 +491,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 genes, tissues, organism, dim_size, 1.0, expected_count
             )
 
-            response = self.app.post("/wmg/v1/query", json=request)
+            response = self.app.post("/wmg/v2/query", json=request)
 
             self.assertEqual(200, response.status_code)
 
@@ -499,7 +499,7 @@ class WmgApiV1Tests(unittest.TestCase):
             self.assertEqual(expected, json.loads(response.data)["term_id_labels"]["cell_types"])
 
     def test__query_empty_request__returns_400(self):
-        response = self.app.post("/wmg/v1/query", json={})
+        response = self.app.post("/wmg/v2/query", json={})
 
         self.assertEqual(400, response.status_code)
         self.assertEqual("'filter' is a required property", json.loads(response.data)["detail"])
@@ -512,7 +512,7 @@ class WmgApiV1Tests(unittest.TestCase):
             ),
         )
 
-        response = self.app.post("/wmg/v1/query", json=request)
+        response = self.app.post("/wmg/v2/query", json=request)
 
         self.assertEqual(400, response.status_code)
 
@@ -524,7 +524,7 @@ class WmgApiV1Tests(unittest.TestCase):
             ),
         )
 
-        response = self.app.post("/wmg/v1/query", json=request)
+        response = self.app.post("/wmg/v2/query", json=request)
 
         self.assertEqual(400, response.status_code)
 
@@ -536,14 +536,14 @@ class WmgApiV1Tests(unittest.TestCase):
             ),
         )
 
-        response = self.app.post("/wmg/v1/query", json=request)
+        response = self.app.post("/wmg/v2/query", json=request)
 
         self.assertEqual(400, response.status_code)
 
-    @patch("backend.wmg.api.v1.fetch_datasets_metadata")
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.fetch_datasets_metadata")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_request_with_filter_dims__returns_valid_filter_dims__base_case(
         self, load_snapshot, ontology_term_label, gene_term_label, fetch_datasets_metadata
     ):
@@ -571,7 +571,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 is_rollup=True,
             )
 
-            response = self.app.post("/wmg/v1/filters", json=filter_0_request)
+            response = self.app.post("/wmg/v2/filters", json=filter_0_request)
 
             expected_filters = {
                 "datasets": [
@@ -595,10 +595,10 @@ class WmgApiV1Tests(unittest.TestCase):
             }
             self.assertEqual(json.loads(response.data)["filter_dims"], expected_filters)
 
-    @patch("backend.wmg.api.v1.fetch_datasets_metadata")
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.fetch_datasets_metadata")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__query_request_with_filter_dims__returns_valid_filter_dims(
         self, load_snapshot, ontology_term_label, gene_term_label, fetch_datasets_metadata
     ):
@@ -659,18 +659,18 @@ class WmgApiV1Tests(unittest.TestCase):
                 filter_0_no_ethnicity_request = dict(filter=filter_0_no_ethnicity_filter, is_rollup=True)
                 # the values for dev_stage terms when a dev stage filter is included should match the values returned
                 # if no filter is passed in for dev stage
-                response = self.app.post("/wmg/v1/filters", json=filter_0_request)
+                response = self.app.post("/wmg/v2/filters", json=filter_0_request)
                 dev_stage_terms = json.loads(response.data)["filter_dims"]["development_stage_terms"]
                 self_reported_ethnicity_terms = json.loads(response.data)["filter_dims"][
                     "self_reported_ethnicity_terms"
                 ]
 
-                no_dev_stage_filter_response = self.app.post("/wmg/v1/filters", json=filter_0_no_dev_stage_request)
+                no_dev_stage_filter_response = self.app.post("/wmg/v2/filters", json=filter_0_no_dev_stage_request)
                 dev_stage_terms_if_no_dev_stage_filters = json.loads(no_dev_stage_filter_response.data)["filter_dims"][
                     "development_stage_terms"
                 ]
 
-                no_ethnicity_filter_response = self.app.post("/wmg/v1/filters", json=filter_0_no_ethnicity_request)
+                no_ethnicity_filter_response = self.app.post("/wmg/v2/filters", json=filter_0_no_ethnicity_request)
                 self_reported_ethnicity_terms_if_no_dev_stage_filters = json.loads(no_ethnicity_filter_response.data)[
                     "filter_dims"
                 ]["self_reported_ethnicity_terms"]
@@ -711,7 +711,7 @@ class WmgApiV1Tests(unittest.TestCase):
                     {"self_reported_ethnicity_ontology_term_id_2": "self_reported_ethnicity_ontology_term_id_2_label"},
                 ]
                 self_reported_ethnicity_0_request = dict(filter=self_reported_ethnicity_0_filter, is_rollup=True)
-                response = self.app.post("/wmg/v1/filters", json=self_reported_ethnicity_0_request)
+                response = self.app.post("/wmg/v2/filters", json=self_reported_ethnicity_0_request)
                 dev_stage_terms = _sort_list_by_dictionary_keys(
                     json.loads(response.data)["filter_dims"]["development_stage_terms"]
                 )
@@ -740,7 +740,7 @@ class WmgApiV1Tests(unittest.TestCase):
                     {"self_reported_ethnicity_ontology_term_id_0": "self_reported_ethnicity_ontology_term_id_0_label"}
                 ]
                 self_reported_ethnicity_1_request = dict(filter=self_reported_ethnicity_1_filter, is_rollup=True)
-                response = self.app.post("/wmg/v1/filters", json=self_reported_ethnicity_1_request)
+                response = self.app.post("/wmg/v2/filters", json=self_reported_ethnicity_1_request)
                 dev_stage_terms_no_dev_filter = _sort_list_by_dictionary_keys(
                     json.loads(response.data)["filter_dims"]["development_stage_terms"]
                 )
@@ -763,7 +763,7 @@ class WmgApiV1Tests(unittest.TestCase):
                     filter=self_reported_ethnicity_1_dev_1_filter, is_rollup=True
                 )
 
-                response = self.app.post("/wmg/v1/filters", json=self_reported_ethnicity_1_dev_1_request)
+                response = self.app.post("/wmg/v2/filters", json=self_reported_ethnicity_1_dev_1_request)
                 dev_stage_terms_dev_filter = _sort_list_by_dictionary_keys(
                     json.loads(response.data)["filter_dims"]["development_stage_terms"]
                 )
@@ -799,7 +799,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 )
                 self_reported_ethnicity_2_request = dict(filter=self_reported_ethnicity_2_filter, is_rollup=True)
                 eth_2_dev_2_request = dict(filter=self_reported_ethnicity_2_dev_2_filter, is_rollup=True)
-                response = self.app.post("/wmg/v1/filters", json=self_reported_ethnicity_2_request)
+                response = self.app.post("/wmg/v2/filters", json=self_reported_ethnicity_2_request)
                 dev_stage_terms_eth_2_no_dev_filter = _sort_list_by_dictionary_keys(
                     json.loads(response.data)["filter_dims"]["development_stage_terms"]
                 )
@@ -809,7 +809,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 self.assertEqual(expected_development_stage_terms, dev_stage_terms_eth_2_no_dev_filter)
                 self.assertEqual(all_self_reported_ethnicity_terms, self_reported_ethnicity_terms_eth_2_no_dev_filter)
 
-                response = self.app.post("/wmg/v1/filters", json=eth_2_dev_2_request)
+                response = self.app.post("/wmg/v2/filters", json=eth_2_dev_2_request)
                 dev_stage_terms_eth_2_dev_2 = _sort_list_by_dictionary_keys(
                     json.loads(response.data)["filter_dims"]["development_stage_terms"]
                 )
@@ -822,9 +822,9 @@ class WmgApiV1Tests(unittest.TestCase):
                 self.assertEqual(dev_stage_terms_eth_2_dev_2, dev_stage_terms_eth_2_no_dev_filter)
                 self.assertNotEqual(eth_stage_terms_eth_2_dev_2, self_reported_ethnicity_terms_eth_2_no_dev_filter)
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__markers_returns_200_and_correct_response(self, load_snapshot, ontology_term_label, gene_term_label):
         with load_realistic_test_snapshot(TEST_SNAPSHOT) as snapshot:
             # setup up API endpoints to use a mocked cube containing all stat values of 1, for a deterministic
@@ -845,7 +845,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 test="ttest",
             )
 
-            response = self.app.post("/wmg/v1/markers", json=request)
+            response = self.app.post("/wmg/v2/markers", json=request)
             received = json.loads(response.data)
             criteria = MarkerGeneQueryCriteria(
                 tissue_ontology_term_id=request["tissue"],
@@ -859,9 +859,9 @@ class WmgApiV1Tests(unittest.TestCase):
             self.assertDictEqual(received, expected)
             self.assertEqual(200, response.status_code)
 
-    @patch("backend.wmg.api.v1.gene_term_label")
-    @patch("backend.wmg.api.v1.ontology_term_label")
-    @patch("backend.wmg.api.v1.load_snapshot")
+    @patch("backend.wmg.api.v2.gene_term_label")
+    @patch("backend.wmg.api.v2.ontology_term_label")
+    @patch("backend.wmg.api.v2.load_snapshot")
     def test__markers_returns_200_and_empty_dictionary_for_bad_celltypes(
         self, load_snapshot, ontology_term_label, gene_term_label
     ):
@@ -884,7 +884,7 @@ class WmgApiV1Tests(unittest.TestCase):
                 test="ttest",
             )
 
-            response = self.app.post("/wmg/v1/markers", json=request)
+            response = self.app.post("/wmg/v2/markers", json=request)
             received = json.loads(response.data)
 
             expected = {"marker_genes": [], "snapshot_id": "realistic-test-snapshot"}
